show images in detail activity

+++ app/src/main/java/br/com/claro/movies/dto/Backdrop.kt
@@ -7,7 +7,7 @@ data class Backdrop(
         @SerializedName("file_path") val filePath: String,
         @SerializedName("height") val height: Int,
         @SerializedName("iso_639_1") val iso6391: String,
-        @SerializedName("vote_average") val voteAverage: Int,
+        @SerializedName("vote_average") val voteAverage: Double,
         @SerializedName("vote_count") val voteCount: Int,
         @SerializedName("width") val width: Int
 )
